Generate be/preg_list.cxx at build time
Use add_custom_command to define a build rule for the source file
instead of generating it ahead of time.  The build rule will be put in
both the be-static and be library targets.  We add an extra dependency
between these targets to serialize them and ensure that the file is
generated at most once.